refactor global package into version package and rename and replace having cli login,logout,validate removed
