+++ cli/config/config.go
@@ -18,7 +18,7 @@
 package config
 
 import (
-	"github.com/arduino/arduino-cli/global"
+	"github.com/arduino/arduino-cli/version"
 	"github.com/spf13/cobra"
 )
