+++ cli/daemon/daemon.go
@@ -23,8 +23,8 @@ import (
 	"net"
 
 	"github.com/arduino/arduino-cli/daemon"
-	"github.com/arduino/arduino-cli/global"
 	"github.com/arduino/arduino-cli/rpc"
+	"github.com/arduino/arduino-cli/version"
 	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
 )
