+++ cli/generatedocs/generatedocs.go
@@ -22,7 +22,7 @@ import (
 	"path/filepath"
 
 	"github.com/arduino/arduino-cli/cli"
-	"github.com/arduino/arduino-cli/global"
+	"github.com/arduino/arduino-cli/version"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/cobra/doc"
