+++ cli/sketch/sketch.go
@@ -18,7 +18,7 @@
 package sketch
 
 import (
-	"github.com/arduino/arduino-cli/global"
+	"github.com/arduino/arduino-cli/version"
 	"github.com/spf13/cobra"
 )
